@@ -1,37 +1,46 @@
 {
-	"name": "next-simple-starter",
-	"version": "1.0.0",
-	"repository": {
-		"url": "https://github.com/mikdatdogru/md-nextjs-boilerplate"
-	},
-	"license": "MIT",
-	"dependencies": {
-		"isomorphic-fetch": "^2.2.1",
-		"next": "^7.0.0",
-		"next-redux-wrapper": "^2.0.0",
-		"react": "^16.6.3",
-		"react-dom": "^16.6.3",
-		"react-redux": "^6.0.0",
-		"redux": "^3.7.2"
-	},
-	"devDependencies": {
-		"babel-eslint": "^10.0.1",
-		"eslint": "^5.12.0",
-		"eslint-config-airbnb": "^17.1.0",
-		"eslint-config-prettier": "^3.3.0",
-		"eslint-plugin-import": "^2.14.0",
-		"eslint-plugin-jsx-a11y": "^6.1.2",
-		"eslint-plugin-prettier": "^3.0.1",
-		"eslint-plugin-react": "^7.12.3",
-		"prettier": "^1.15.3",
-		"workbox-webpack-plugin": "^3.4.1"
-	},
-	"scripts": {
-		"dev": "node server",
-		"build": "next build",
-		"start": "NODE_ENV=production node server",
-		"lint": "eslint src",
-		"lint:fix": "eslint . --fix",
-		"prettier": "prettier --config .prettierrc --write \"./**/*.js\""
-	}
+  "name": "md-nextjs-boilerplate",
+  "version": "1.0.0",
+  "repository": {
+    "url": "https://github.com/mikdatdogru/md-nextjs-boilerplate"
+  },
+  "license": "MIT",
+  "dependencies": {
+    "isomorphic-fetch": "^2.2.1",
+    "next": "^8.1.0",
+    "next-offline": "^4.0.1",
+    "next-redux-wrapper": "^3.0.0-alpha.2",
+    "react": "^16.8.6",
+    "react-dom": "^16.8.6",
+    "react-redux": "^7.0.3",
+    "redux": "^4.0.1"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.4.5",
+    "@storybook/addon-actions": "^5.0.11",
+    "@storybook/addon-links": "^5.0.11",
+    "@storybook/addons": "^5.0.11",
+    "@storybook/react": "^5.0.11",
+    "babel-eslint": "^10.0.1",
+    "babel-loader": "^8.0.6",
+    "eslint": "^5.16.0",
+    "eslint-config-airbnb": "^17.1.0",
+    "eslint-config-prettier": "^4.3.0",
+    "eslint-plugin-import": "^2.17.3",
+    "eslint-plugin-jsx-a11y": "^6.2.1",
+    "eslint-plugin-prettier": "^3.1.0",
+    "eslint-plugin-react": "^7.13.0",
+    "prettier": "^1.17.1",
+    "workbox-webpack-plugin": "^4.3.1"
+  },
+  "scripts": {
+    "dev": "node server",
+    "build": "next build",
+    "start": "NODE_ENV=production node server",
+    "lint": "eslint src",
+    "lint:fix": "eslint . --fix",
+    "prettier": "prettier --config .prettierrc --write \"./**/*.js\"",
+    "storybook": "start-storybook -p 6006",
+    "build-storybook": "build-storybook"
+  }
 }
